Add files via upload
Restructured project directories; removed nextfloor() method from the CollisionDetector class: the method is now useless due to changes in the mechanics of gravity; removed MapElement class and relative subclasses due to rethinking of the game map structure; added Map class containing methods for random map generation, the game keeps going on indefinitely, and a new map is generated when the player reaches the end of the current one. Added RandomGenerator class, containing a method to generate a random number in a given range, which will have many use cases in map generation and computing spawn coordinates for enemies.